+++ tests/visitors/jinja_visitor/conftest.py
@@ -1,3 +1,2 @@
 import pytest
-
 from mau.visitors.jinja_visitor import JinjaVisitor